requirements.txt
joblib==0.14.1
Flask==1.1.1
numpy==1.18.2
requests==2.20.0
beautifulsoup4==4.9.1
gunicorn==20.0.4
lxml==4.2.5
scikit_learn==0.23.2
yagmail==0.11.224
psycopg2-binary==2.8.5